pyproject.toml
[project]
name = "Alexion-demo-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.13",
    "chromadb>=0.5.0",
    "ollama>=0.2.0",
    "anthropic>=0.28.0",
    "openai>=1.30.0",
    "pyvis>=0.3.2",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
]